Initial stab at test suite
Initial test cases that cover the broadcasting, receiving,
and sortedness of tweets. The specs are doing some legwork to
avoid conflicts over the shared state of the genserver- those arise
due to the concurrent and random execution order of tests in Phoenix.
Definitely opportunities for improvement and my education here