Support updating tweets with new popularity values
Previously Phoenix would only delete a tweet if it weren't in the
top 50. Now, we check if we are currently tracking the tweet in our
list, rather than inserting blindly. If it's being tracked, we
delete first then insert the updated one. The server side code
passes to the JS an extra index indicating that this tweet is an
update, not a new tweet.